+++ examples/fast_sym.cxx
@@ -39,9 +39,9 @@ int fast_sym(int const     n,
   CTF_Vector Cs(n, ctf);
 
   {
-    int64_t * indices;
+    long long * indices;
     double * values;
-    int64_t size;
+    long long size;
     srand48(173*rank);
 
     A.get_local_data(&size, &indices, &values);
